Changed Level Layouts for level one and 2 slightly and more
Added more things that would make levels easier to traverse and then added a level transition (temporary implementation) that leads into level 2 from level one. There's one also in level 2 that's supposed to lead back to the Main Menu but I can't get it to work for some reason. I also disabled level 2 battle encounters as it led to crashes so that'll also be gone temporarily. The level transition and main menu transition scripts are "LevelTransition" and "ExitTransition"